+++ src/components/App.js
@@ -3,14 +3,12 @@ import { Switch, BrowserRouter, Route } from 'react-router-dom';
 
 // pages
 import Home from '../pages/home';
-import GiphyPage from '../pages/giphy-page';
 
 function App() {
   return (
     <BrowserRouter>
       <Switch>
         <Route exact path="/" component={Home} />
-        <Route exact path="/giph" component={GiphyPage} />
       </Switch>
     </BrowserRouter>
   );
